refactor: rename _git_utils.py to _git_diff_utils.py for specificity
